Fix: branch by abstraction
